Move exceptions to exceptions submodule
